backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
aiofiles==23.2.1
requests==2.32.3
google-auth==2.30.0
PyJWT==2.8.0
